% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAlpha.R
\name{calculateAlpha}
\alias{calculateAlpha}
\title{Calculate alpha (relationship between MSD and normal diffusion)}
\usage{
calculateAlpha(alphaMat, tstep)
}
\arguments{
\item{alphaMat}{matrix of msd curves, each col is a track, each row is time lag (will contain NAs)}

\item{tstep}{variable. Time step in seconds}
}
\value{
data frame
}
\description{
Normal diffusion is alpha = 1. Subdiffusion is alpha < 1 and superdiffusion is alpha > 1.
Input is a data matrix of msd curves.
Output is mean of log2(alpha), one value for each trace.
}
