% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCVE.R
\name{calculateCVE}
\alias{calculateCVE}
\title{Calculate CVE (covariance-based estimator)}
\usage{
calculateCVE(xMat, yMat, tlist, tstep)
}
\arguments{
\item{xMat}{matrix of x displacements, each col is a track, each row is time (will contain NAs)}

\item{yMat}{matrix of y displacements, each col is a track, each row is time (will contain NAs)}

\item{tlist}{list of trace names}

\item{tstep}{variable. Time step in seconds}
}
\value{
data frame
}
\description{
From Vestergaard et al., Physical Review E (2019) 89, 022726
Input is two data matrices of X and Y displacements for each trace.
Output is estimator of D and sigma for each trace.
}
