% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFD.R
\name{calculateFD}
\alias{calculateFD}
\title{Calculate fractal dimension (FD)}
\usage{
calculateFD(dataList)
}
\arguments{
\item{dataList}{list of a data frame (must include at a minimum - trace (track ID), x, y and frame (in real coords)) and a calibration data frame}
}
\value{
data frame
}
\description{
Calculate for each track, its fractal dimension
Katz & George (1985) define fractal dimension (D) as log(n) / (log(n) + log(d/L))
where n is the number of steps, d is the longest of all possible point-to-point distances and
L is the cumulative length of the track.
D is ~1 for directed trajectories, ~2 for confined and ~3 for subdiffusion
Here we calculate this and store D (called fd) and d (called wide) for return.
Note that this method does not take into account gaps in the track.
For a track with many gaps, n will be lowered.
}
\examples{
xmlPath <- system.file("extdata", "ExampleTrackMateData.xml", package="TrackMateR")
tmObj <- readTrackMateXML(XMLpath = xmlPath)
tmObj <- correctTrackMateData(dataList = tmObj, xyscalar = 0.04)
fdDF <- calculateFD(dataList = tmObj)
}
