% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateJD.R
\name{calculateJD}
\alias{calculateJD}
\title{Calculate Jump Distance (JD)}
\usage{
calculateJD(df, deltaT = 1)
}
\arguments{
\item{df}{data frame must include at a minimum - trace (track ID), x, y and t (in real coords)}

\item{deltaT}{integer to represent the multiple of frames that are to be analysed}
}
\value{
vector of jump distances, NAs removed
}
\description{
Calculation of the JD of multiple tracks.
Calculation is equivalent to a single time lag point on the ensemble MSD curve, typically represented as a histogram
Input is a data frame of tracks imported using readTrackMateXML()
The default time step is one frame - which is the equivalent to the plot generated to show displacement versus time.
}
\examples{
xmlPath <- "~/Desktop/FakeTracks.xml"
datalist <- readTrackMateXML(XMLpath = xmlPath)
data <-  datalist[[1]]
data <- correctTrackMateData(data, xy = 0.04)
jdvec <- calculateJD(data, deltaT = 2)
}
