% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMSD.R
\name{calculateMSD}
\alias{calculateMSD}
\title{Calculate Mean Squared Displacement (MSD)}
\usage{
calculateMSD(df, N = 4, short = 0)
}
\arguments{
\item{df}{data frame must include at a minimum - trace (track ID), x, y and t (in real coords)}

\item{N}{numeric variable for MSD. dt should be up to 1/N of number of data points (4 recommended)}

\item{short}{numeric variable for the shortest number of points we will analyse. Note, this uses the number of frames from start, not number of points in track, i.e. a track with <short points and many gaps will remain}
}
\value{
data frame
}
\description{
Input is a data frame of tracks imported using readTrackMateXML()
}
\examples{
xmlPath <- "~/Desktop/FakeTracks.xml"
data <- readTrackMateXML(XMLpath = xmlPath)
data <- correctTrackMateData(data, xy = 0.04)
msdDF <- calculateMSD(data, N = 3, short = 8)
}
