% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_functions.R
\name{compareGTDatasets}
\alias{compareGTDatasets}
\title{Compare ground truth datasets}
\usage{
compareGTDatasets(
  path = NULL,
  xyscale = 0.04,
  xyunit = "um",
  tscale = 0.06,
  tunit = "s",
  ...
)
}
\arguments{
\item{path}{string, filepath to directory equivalent to Data in compareDatasets()}

\item{xyscale}{numeric, pixel size (ground truth data is in pixel units)}

\item{xyunit}{string, spatial units}

\item{tscale}{numeric, frame interval (ground truth data is in frames)}

\item{tunit}{string, temporal units}

\item{...}{pass additional parameters to modify the defaults (N, short, deltaT, mode, nPop, init, timeRes, breaks, radius)}
}
\value{
multiple pdf reports
}
\description{
Requires Ground Truth csv files to be organised into subfolders named according to the condition.
Outputs are saved to `Output/Plots/GT/` in the working directory.
}
