% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTrackMateXML2CSV.R
\name{convert_trackmate_xml_to_csv}
\alias{convert_trackmate_xml_to_csv}
\title{Convert TrackMate XML file to CSV format}
\usage{
convert_trackmate_xml_to_csv(
  path = NULL,
  min_points = 10,
  input_dir = "Data/",
  output_dir = "Output/Data/",
  slim = TRUE,
  columns = c("trace", "frame", "y", "x"),
  pixels = FALSE,
  no_save = FALSE
)
}
\arguments{
\item{path}{path to the TrackMate XML file}

\item{min_points}{minimum number of points in a trace to be included in the
output csv file (default: 10). Set to 0 to include all traces.}

\item{input_dir}{directory where the input XML files are located (default:
"Data/")}

\item{output_dir}{directory where the output csv file will be saved (default:
"Output/Data/").}

\item{slim}{logical, if TRUE, only the minimum data required to process the
tracks is read (default: TRUE)}

\item{columns}{columns to be included in the output csv file (default: c("
trace", "frame", "y", "x")). Set to NULL to include all columns from the
TrackMate XML file.}

\item{pixels}{logical, if TRUE, x and y will be converted to pixels using the
calibration information from the TrackMate XML file (default: FALSE)}

\item{no_save}{logical, if TRUE, the output csv file will not be saved and a
data frame returned to the user (default: FALSE)}
}
\value{
saves a csv file
}
\description{
This function reads a TrackMate XML file and save it as a csv
  file for use in other contexts.

  There are options to:
- filter out traces with less than n points
- save in a location other than Output/Data
- select columns for export
- x and y can be scaled (default) or converted to pixels

  The typical scenario is that the XML file is in `input_dir` and the
  converted csv will be saved to `output_dir`. Any subfolders below
  `input_dir` will be recreated in `output_dir`. If a full path is supplied
  as the `path`, the `input_dir` should be adjusted to be the directory
  containing the XML file or containing the subdirectory containing the XML
  file.

  Note: the calibration of the input TrackMate XML file must be correct.
}
\examples{
\dontrun{
files <- list.files(path = "Data", pattern = "\\\\.xml$", full.names = TRUE)
lapply(files, convert_trackmate_xml_to_csv)
}
}
