% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctTrackMateData.R
\name{correctTrackMateData}
\alias{correctTrackMateData}
\title{Correct distance and time of imported TrackMate data.}
\usage{
correctTrackMateData(df, xyscalar = 1, tscalar = 1)
}
\arguments{
\item{df}{data frame of imported track mate data}

\item{xyscalar}{numeric multiplier to correct pixel size of original movie. Assumes isotropic scaling, i.e. pixel height = pixel width}

\item{tscalar}{numeric multiplier to correct frame interval of original movie. Frame interval of tracked data.}
}
\value{
data frame
}
\description{
If the TrackMate data is in pixels and/or frames, the data frame can be converted with this function.
}
\examples{
xmlPath <- "~/Desktop/FakeTracks.xml"
datalist <- readTrackMateXML(XMLpath = xmlPath)
data <-  datalist[[1]]
# in the case where pixel size is 0.03 um and original data is 1 pixel, xyscalar = 0.03
data <- correctTrackMateData(df = data, xyscalar = 0.03)
}
