% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTrackDensity.R
\name{find_distances}
\alias{find_distances}
\title{Find distance between xy coordinate and a series of other xy coordinates}
\usage{
find_distances(xx, yy, df)
}
\arguments{
\item{xx}{x coord of point for comparison}

\item{yy}{y coord of point for comparison}

\item{df}{data frame containing x and y columns for other points}
}
\value{
numeric vector of distances
}
\description{
Find distance between xy coordinate and a series of other xy coordinates
}
