% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTrackDensity.R
\name{find_td_A1}
\alias{find_td_A1}
\title{Track Density - Find area A1}
\usage{
find_td_A1(x, r)
}
\arguments{
\item{x}{value}

\item{r}{radius}
}
\value{
numeric variable
}
\description{
Track Density - Find area A1
}
\keyword{internal}
