% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTrackDensity.R
\name{find_td_A2}
\alias{find_td_A2}
\title{Track Density - Find area A2}
\usage{
find_td_A2(x, y, r)
}
\arguments{
\item{x}{value}

\item{y}{value}

\item{r}{radius}
}
\value{
numeric variable
}
\description{
Track Density - Find area A2
}
\keyword{internal}
