% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_misc.R
\name{findLog2YAxisLimits}
\alias{findLog2YAxisLimits}
\title{Find log2 y limits for symmetrical axis}
\usage{
findLog2YAxisLimits(x)
}
\arguments{
\item{input}{vector of log2 values}
}
\value{
list two limits for y-axis
}
\description{
Find log2 y limits for symmetrical axis
}
\keyword{internal}
