% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeComparison.R
\name{makeComparison}
\alias{makeComparison}
\title{Make Comparison Plots}
\usage{
makeComparison(
  df,
  msddf,
  units = c("um", "s"),
  msdplot = "linlin",
  titleStr = "Comparison",
  subStr = NULL
)
}
\arguments{
\item{df}{data frame called megareport}

\item{msddf}{data frame of msd averages per dataset}

\item{units}{character vector of space and time units for axes}

\item{msdplot}{keyword used to specify msdplot axes, either "loglog", "linlin" (default), "loglin" or "linlog"}

\item{titleStr}{string used as the title for the report}

\item{subStr}{string used as the subtitle for the report}
}
\value{
patchwork ggplot
}
\description{
A series of ggplots to compare between conditions.
Called from `compareDatasets()` this function generates plots using summary data of datasets, per condition.
}
