% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSummary.R
\name{makeSummary}
\alias{makeSummary}
\title{Make Summary}
\usage{
makeSummary(
  df,
  msdlist,
  jumpdata,
  jumptime,
  units = c("um", "s"),
  titleStr = "Summary",
  subStr = NULL
)
}
\arguments{
\item{df}{imported TrackMate data with correct units}

\item{msdlist}{list of 2 data frames: MSD summary and alpha summary = collated outputs from calculateMSD()}

\item{jumpdata}{data frame of jump data}

\item{jumptime}{variable to be passed to timeRes}

\item{units}{character vector to describe units (defaults are um, micrometres and  s, seconds)}

\item{titleStr}{string used as the title for the summary}

\item{subStr}{string used as the subtitle for the summary}
}
\value{
patchwork ggplot
}
\description{
Generate several plots to visualise TrackMate data and generate a report.
Note that the units are hard-coded as um and s.
The use of um is because ggsave does not currently save unicode to PDF reliably.
}
