% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSummaryReport.R
\name{makeSummaryReport}
\alias{makeSummaryReport}
\title{Make Summary Report}
\usage{
makeSummaryReport(
  tmList,
  msdList,
  jumpList,
  tddf,
  fddf,
  titleStr = "",
  subStr = "",
  auto = FALSE,
  summary = FALSE,
  ...
)
}
\arguments{
\item{tmList}{list of trackmate data and calibration}

\item{msdList}{MSD summary and alpha list = output from calculateMSD()}

\item{jumpList}{list of a data frame of jump data and a variable to be passed to timeRes}

\item{tddf}{data frame of track density data}

\item{fddf}{data frame of fractal dimension data}

\item{titleStr}{string used as the title for the report}

\item{subStr}{string used as the subtitle for the report}

\item{auto}{boolean which selects for returning the patchwork report (FALSE) or a list of the patchwork report and a data frame of summary (TRUE)}

\item{summary}{boolean which selects for report (FALSE) or summary (TRUE)}

\item{...}{additional arguments passed to methods}
}
\value{
patchwork ggplot or a list of patchwork ggplot and data frame of summary data
}
\description{
Generate several plots to visualise TrackMate data and generate a report.
The use of um is because ggsave does not currently save unicode to PDF reliably.
The code is switchable to accommodate making a "report" (one dataset) or a "summary" (several related datasets combined)
}
\examples{
xmlPath <- system.file("extdata", "ExampleTrackMateData.xml", package="TrackMateR")
tmObj <- readTrackMateXML(XMLpath = xmlPath)
tmObj <- correctTrackMateData(tmObj, xyscalar = 0.04)
tmDF <- tmObj[[1]]
calibrationDF <- tmObj[[2]]
msdObj <- calculateMSD(df = tmDF, method = "ensemble", N = 3, short = 8)
jdObj <- calculateJD(dataList = tmObj, deltaT = 1)
tdDF <- calculateTrackDensity(dataList = tmObj, radius = 1.5)
fdDF <- calculateFD(dataList = tmObj)
fileName <- tools::file_path_sans_ext(basename(xmlPath))
reportObj <- makeSummaryReport(tmList = tmObj, msdList = msdObj, jumpList = jdObj,
tddf = tdDF, fddf = fdDF, titleStr = "Report", subStr = fileName, auto = TRUE)
}
