% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_misc.R
\name{mergeDataFramesForExport}
\alias{mergeDataFramesForExport}
\title{Merge Data Frames for export}
\usage{
mergeDataFramesForExport(x, y)
}
\arguments{
\item{x}{data frame to be coerced}

\item{y}{data frame to be coerced}
}
\description{
Properties for individual traces (tracks) from different experiments need to be merged for export.
Utility function, not for export.
}
\keyword{internal}
