% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_plots.R
\name{plot_tm_allTracks}
\alias{plot_tm_allTracks}
\title{Make a plot of all tracks}
\usage{
plot_tm_allTracks(
  df,
  summary = FALSE,
  xstr = "x",
  ystr = "y",
  alphaLevel = 0.5
)
}
\arguments{
\item{df}{data frame of TrackMate data}

\item{summary}{boolean to specify if plot is of one dataset or several related datasets}

\item{xstr}{string to label x-axis}

\item{ystr}{string to label y-axis}

\item{alphaLevel}{numeric variable to set alpha for the plot}
}
\value{
ggplot
}
\description{
Make a plot of all tracks
}
