% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_plots.R
\name{plot_tm_dee}
\alias{plot_tm_dee}
\title{Make a histogram of D values}
\usage{
plot_tm_dee(df, median_dee = NULL, xstr = "D", ystr = "Frequency")
}
\arguments{
\item{df}{data frame of D values}

\item{median_dee}{variable for adding label to plot}

\item{xstr}{string to label x-axis}

\item{ystr}{string to label y-axis}
}
\value{
ggplot
}
\description{
Make a histogram of D values
}
