% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_plots.R
\name{plot_tm_fd}
\alias{plot_tm_fd}
\title{Make a histogram of fractal dimension}
\usage{
plot_tm_fd(df, auto = FALSE)
}
\arguments{
\item{df}{data frame of fractal dimension data}

\item{auto}{boolean to switch between returning a ggplot and a list of ggplot and variable}
}
\value{
ggplot or list of ggplot and variable
}
\description{
Make a histogram of fractal dimension
}
