% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_plots.R
\name{plot_tm_intensityHist}
\alias{plot_tm_intensityHist}
\title{Make a histogram of intensities}
\usage{
plot_tm_intensityHist(
  input,
  summary = FALSE,
  xstr = NULL,
  ystr = NULL,
  auto = FALSE
)
}
\arguments{
\item{input}{either a data frame of TrackMate data or list of TrackMate data frame and calibration data frame}

\item{summary}{boolean to specify if plot is of one dataset or several related datasets}

\item{xstr}{string to label x-axis}

\item{ystr}{string to label y-axis}

\item{auto}{boolean to switch between returning a ggplot and a list of ggplot and variable}
}
\value{
ggplot or list of ggplot and variable
}
\description{
Function uses the maximum value of the mean_intensity column for all traces in supplied TrackMate data
}
