% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_plots.R
\name{plot_tm_NMSD}
\alias{plot_tm_NMSD}
\title{Plot several (n) MSD curves}
\usage{
plot_tm_NMSD(df, xlog = FALSE, ylog = FALSE, auto = FALSE)
}
\arguments{
\item{df}{dataframe of MSD summary data from multiple datasets (labelled by dataid)}

\item{xlog}{boolean to request log10 x axis}

\item{ylog}{boolean to request log10 y axis}

\item{auto}{boolean to request plot only, TRUE gives plot and summary dataframe as a list}
}
\value{
ggplot or list of ggplot and dataframe of summary data
}
\description{
Generate a plot of several MSD curves together with a summary curve.
This function is used to compile multiple datasets from the same condition.
}
