% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_plots.R
\name{plot_tm_width}
\alias{plot_tm_width}
\title{Make a histogram of the largest point-to-point distance in each track}
\usage{
plot_tm_width(df, units = c("um", "s"), auto = FALSE)
}
\arguments{
\item{df}{data frame of fractal dimension data}

\item{units}{character vector of space and time units (default is um and s)}

\item{auto}{boolean to switch between returning a ggplot and a list of ggplot and variable}
}
\value{
ggplot or list of ggplot and variable
}
\description{
Make a histogram of the largest point-to-point distance in each track
}
