% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMSD.r
\name{plotMSD}
\alias{plotMSD}
\title{Make a plot of MSD data}
\usage{
plotMSD(df, units = "s", xlog = FALSE, ylog = FALSE)
}
\arguments{
\item{df}{MSD summary = output from calculateMSD()}

\item{units}{string to describe time units (default is s, seconds)}

\item{xlog}{boolean to request log10 x axis}

\item{ylog}{boolean to request log10 y axis}
}
\value{
S3 ggplot
}
\description{
Input is the output from CalculateMSD()
}
