% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNMSD.R
\name{plotNMSD}
\alias{plotNMSD}
\title{Plot several (n) MSD curves}
\usage{
plotNMSD(df)
}
\arguments{
\item{df}{dataframe of MSD summary data from multiple datasets (labelled by dataid)}
}
\value{
ggplot
}
\description{
Generate a plot of several MSD curves together with a summary curve.
}
