% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_misc.R
\name{processEllipsis}
\alias{processEllipsis}
\title{Process additional parameters}
\usage{
processEllipsis(input)
}
\arguments{
\item{input}{list of ellipsis arguments}
}
\value{
list of arguments adjusted for defaults
}
\description{
Ensure ellipsis parameters have the default values
This function is used in two separate workflows, so a single function to edit makes sense.
}
\keyword{internal}
