% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrackMateXML.r
\name{readTrackMateXML}
\alias{readTrackMateXML}
\title{Read TrackMate XML output files.}
\usage{
readTrackMateXML(XMLpath)
}
\arguments{
\item{XMLpath}{path to the xml file}
}
\value{
data frame
}
\description{
Produces a data frame of all spots from filtered tracks, ordered by track number.
}
\examples{
xmlPath <- "~/Desktop/FakeTracks.xml"
df <- readTrackMateXML(XMLpath = xmlPath)
}
