% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportDataset.R
\name{reportDataset}
\alias{reportDataset}
\title{Report Dataset}
\usage{
reportDataset(tmList, ...)
}
\arguments{
\item{tmList}{list of TrackMate data and calibration}

\item{...}{pass additional parameters to modify the defaults (N, short, deltaT, mode, nPop, init, timeRes, breaks, radius)}
}
\value{
patchwork ggplot
}
\description{
This function automates the process of making a report.
If more control is needed over the default parameters:
a user can generate MSD, jump distance and track density objects from their imported TrackMate data, and make a report manually.
}
