% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_misc.R
\name{setupOutputPath}
\alias{setupOutputPath}
\title{Setup Output Path}
\usage{
setupOutputPath(fpath)
}
\arguments{
\item{fpath}{string containing a path that may or may not exist}
}
\description{
In the working directory, we need to check if a path to a directory exists, and if not create it.
Ensures smooth running of the automated functions in TestTrackMateR.
}
