% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateJD.R
\name{calculateJD}
\alias{calculateJD}
\title{Calculate Jump Distance (JD)}
\usage{
calculateJD(dataList, deltaT = 1)
}
\arguments{
\item{dataList}{list of data frame (must include at a minimum - trace (track ID), x, y and t (in real coords)) and calibration}

\item{deltaT}{integer to represent the multiple of frames that are to be analysed}
}
\value{
a list of data frame of jump distances, NAs removed; and a numeric variable (jumptime)
}
\description{
Calculation of the JD of multiple tracks.
Calculation is equivalent to a single time lag point on the ensemble MSD curve, typically represented as a histogram
Input is a data frame of tracks imported using readTrackMateXML()
The default time step is one frame - which is the equivalent to the plot generated to show displacement versus time.
}
\examples{
xmlPath <- system.file("extdata", "ExampleTrackMateData.xml", package="TrackMateR")
tmObj <- readTrackMateXML(XMLpath = xmlPath)
tmObj <- correctTrackMateData(tmObj, xyscalar = 0.04)
jdObj <- calculateJD(dataList = tmObj, deltaT = 2)
}
