% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDatasets.R
\name{compareDatasets}
\alias{compareDatasets}
\title{Compare datasets}
\usage{
compareDatasets()
}
\value{
multiple pdf reports
}
\description{
Requires TrackMate XML files to be organised into subfolders named according to the condition.
If these condition folders are in `Data/` within the working directory, the code will run automatically.
Since there is no easy cross-platform way for the user to interactively pick a directory, the organisation of files in `Data/` is a requirement.
The `Data/` folder can be elsewhere on your computer, just change the `wd` prior to running `compareDatasets()` and the routine will run.
The code will process all the datasets individually, compile them according to condition and compare across conditions.
Outputs are saved to `Output/Plots/` in the working directory.
}
\details{
If TrackMate XML files require recalibration, this is possible by placing a csv file into each subfolder.
All xml files in that folder whose calibration does not match the calibration csv file will be altered.
Ideally, all conditions should have the same scaling, and within a condition they should be similar.
The code will run if this is not the case, but beware that these discrepancies are not detected.
For example, comparing two datasets in um/s with one in mm/min.
}
