% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctTrackMateData.R
\name{correctTrackMateData}
\alias{correctTrackMateData}
\title{Correct distance and time of imported TrackMate data.}
\usage{
correctTrackMateData(df, xysize = 1, tsize = 1)
}
\arguments{
\item{df}{data frame of imported track mate data}

\item{xysize}{pixel size of original movie. Assumes isotropic scaling, i.e. pixel height = pixel width}

\item{tsize}{time. Frame interval of tracked data.}
}
\value{
data frame
}
\description{
If the TrackMate data is in pixels and/or frames, the data frame can be converted with this function.
}
\examples{
xmlPath <- "~/Desktop/FakeTracks.xml"
data <- readTrackMateXML(XMLpath = xmlPath)
data <- correctTrackMateData(data, xy = 0.03)
}
