% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctTrackMateData.R
\name{correctTrackMateData}
\alias{correctTrackMateData}
\title{Correct distance and time of imported TrackMate data.}
\usage{
correctTrackMateData(
  dataList,
  xyscalar = 1,
  tscalar = 1,
  xyunit = NULL,
  tunit = NULL
)
}
\arguments{
\item{dataList}{a list of a data frame (of track data) and a calibration data frame (from TrackMate XML)}

\item{xyscalar}{numeric multiplier to correct pixel size of original movie. Assumes isotropic scaling, i.e. pixel height = pixel width}

\item{tscalar}{numeric multiplier to correct frame interval of original movie. Frame interval of tracked data.}

\item{xyunit}{string to describe spatial units}

\item{tunit}{string to describe temporal unit}
}
\value{
list of two data frames
}
\description{
If the TrackMate data is in pixels and/or frames, the data frame can be converted with this function.
}
\examples{
xmlPath <- system.file("extdata", "ExampleTrackMateData.xml", package="TrackMateR")
tmObj <- readTrackMateXML(XMLpath = xmlPath)
# in the case where pixel size is 0.03 um and original data is 1 pixel, xyscalar = 0.03
tmObj <- correctTrackMateData(dataList = tmObj, xyscalar = 0.03)
}
