% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReport.R
\name{makeReport}
\alias{makeReport}
\title{Make Report}
\usage{
makeReport(df, msdlist, titleStr, subStr)
}
\arguments{
\item{df}{imported TrackMate data with correct units}

\item{msdlist}{MSD summary and alpha list = output from calculateMSD()}

\item{titleStr}{string used as the title for the report}

\item{subStr}{string used as the subtitle for the report}
}
\value{
patchwork ggplot
}
\description{
Generate several plots to visualise TrackMate data and generate a report.
Note that the units are hard-coded as um and s.
The use of um is because ggsave does not currently save unicode to PDF reliably.
}
\examples{
xmlPath <- "~/Desktop/FakeTracks.xml"
data <- readTrackMateXML(XMLpath = xmlPath)
data <- correctTrackMateData(data, xy = 0.04)
result <- calculateMSD(data, method = "ensemble", N = 3, short = 8)
fileName <- tools::file_path_sans_ext(basename(xmlPath))
makeReport(data, result, "", fileName)
}
