% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReport.R
\name{makeReport}
\alias{makeReport}
\title{Make Report}
\usage{
makeReport(df, msddf, titleStr)
}
\arguments{
\item{df}{imported TrackMate data with correct units}

\item{msddf}{MSD summary = output from calculateMSD()}

\item{titleStr}{string used as the title for the report}
}
\value{
patchwork ggplot
}
\description{
Generate several plots to visualise TrackMate data and generate a report.
}
\examples{
xmlPath <- "~/Desktop/FakeTracks.xml"
data <- readTrackMateXML(XMLpath = xmlPath)
data <- correctTrackMateData(data, xy = 0.04)
msdDF <- calculateMSD(data, method = "ensemble", N = 3, short = 8)
fileName <- tools::file_path_sans_ext(xmlPath)
makeReport(data, msdDF, fileName)
}
