% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_plots.R
\name{plot_tm_alpha}
\alias{plot_tm_alpha}
\title{Make a histogram of alpha values}
\usage{
plot_tm_alpha(df, median_alpha = NULL, xstr = NULL, ystr = NULL)
}
\arguments{
\item{df}{data frame of TrackMate data}

\item{median_alpha}{variable for adding label to plot}

\item{xstr}{string to label x-axis}

\item{ystr}{string to label y-axis}
}
\value{
ggplot
}
\description{
Make a histogram of alpha values
}
