% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_plots.R
\name{plot_tm_neighbours}
\alias{plot_tm_neighbours}
\title{Make a histogram of track density (number of neighbours)}
\usage{
plot_tm_neighbours(df)
}
\arguments{
\item{df}{data frame of TrackMate data}
}
\value{
ggplot
}
\description{
Make a histogram of track density (number of neighbours)
}
