% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_plots.R
\name{plot_tm_NMSD}
\alias{plot_tm_NMSD}
\title{Plot several (n) MSD curves}
\usage{
plot_tm_NMSD(df)
}
\arguments{
\item{df}{dataframe of MSD summary data from multiple datasets (labelled by dataid)}
}
\value{
ggplot
}
\description{
Generate a plot of several MSD curves together with a summary curve.
This function is used to compile multiple datasets from the same condition.
}
