% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMSD.r
\name{plotMSD}
\alias{plotMSD}
\title{Make a plot of MSD data}
\usage{
plotMSD(
  df,
  units = c("um", "s"),
  bars = FALSE,
  xlog = FALSE,
  ylog = FALSE,
  auto = FALSE
)
}
\arguments{
\item{df}{MSD summary = output from calculateMSD()}

\item{units}{character vector to describe units (defaults are um, micrometres and  s, seconds)}

\item{bars}{boolean to request error bars (1 x SD)}

\item{xlog}{boolean to request log10 x axis}

\item{ylog}{boolean to request log10 y axis}

\item{auto}{boolean to request plot only, TRUE gives plot and D as a list}
}
\value{
ggplot or ggplot and variable
}
\description{
Generate a plot of MSD over a series of increasing time lags.
Input is the output from CalculateMSD(), so the plot will display the ensemble or time-averaged MSD (whatever was requested)
A fit to the first four points is displayed to evaluate alpha. Diffusion coefficient from this fit is displayed top-left.
}
\examples{
xmlPath <- "~/Desktop/FakeTracks.xml"
datalist <- readTrackMateXML(XMLpath = xmlPath)
data <-  datalist[[1]]
# use the ensemble method and only look at tracks with more than 8 points
msdobj <- calculateMSD(df = data, method = "ensemble", N = 3, short = 8)
msddf <- msdobj[[1]]
plotMSD(msddf, bars = FALSE)
}
