% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrackMateXML.r
\name{readTrackMateXML}
\alias{readTrackMateXML}
\title{Read TrackMate XML output files.}
\usage{
readTrackMateXML(XMLpath)
}
\arguments{
\item{XMLpath}{path to the xml file}
}
\value{
list of two data frames
}
\description{
Produces a data frame of all spots from filtered tracks, ordered by track number.
A warning is generated if the scaling is in pixels rather than real units.
}
\examples{
xmlPath <- "~/Desktop/FakeTracks.xml"
tmObj <- readTrackMateXML(XMLpath = xmlPath)
# get the track data in a data frame
tmDF <-  tmObj[[1]]
# get the calibration data in a data frame
calibrationDF <- tmObj[[2]]
}
